fix(lnn): Updated `Cell` implementation to fix training bug.
Originally, only return the hidden state without a prediction which prevents models from training.

To account for this problem, added a projection layer with a residual connection so that the cell can learn the difference between the data and the hidden state.

Now returns both the residual output and the hidden state.

Additionally, correct head connections to 'sparse' ones for consistency.