refactor(config): Simplified agent `config` creation.
Creating a config model used to require several steps:
- Init 'ModelDetails' and 'TorchConfig'
- Set 'env' and 'TrainConfig' in train method
- Use 'config()' method to merge them

It's a long winded process that shouldn't be that complicated.

Now, we init the config with most of the details (except 'env' and 'TrainConfig') and create a new config model instance using 'copy' when running the 'train' method.

The config model is now directly attached to the agent and is simpler to call and update.