feat(metrics): Moved `TrainMetrics` into`TrainState` for callbacks.
With 'real-time' analytics, metrics need to be interacted with every episode. Logic will need to be added to new callbacks that handles this.

To prepare for this, we move 'TrainMetrics' into the 'TrainState'. This also simplifies the codebase slightly, such as 'avg_reward' in 'EarlyStopping' callback can now be called directly rather than passed in.