refactor(metrics): Updated training metrics name for clarity.
Models that use the 'ReplayBuffer' can be classed as 'episodic' models that focus on using episodes with timesteps.

Models that use the 'RolloutBuffer' are 'timestep' models that focus only on timesteps.

For simplicity, separation of concerns and backwards compatibility we need separate 'TrainMetrics' and 'TrainHandlers' for each type. Current type is 'Episode' related. For clarity, changed 'TrainMetrics' -> 'EpisodeTrainMetrics'.